requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
appium-python-client==1.1.0
    # via -r requirements.in
astroid==2.5.6
    # via
    #   -r requirements.in
    #   pylint
atomicwrites==1.4.0
    # via
    #   -r requirements.in
    #   pytest
attrs==21.2.0
    # via
    #   -r requirements.in
    #   pytest
click==8.0.1
    # via
    #   -r requirements.in
    #   pip-tools
colorama==0.4.4
    # via
    #   -r requirements.in
    #   click
    #   pylint
    #   pytest
iniconfig==1.1.1
    # via
    #   -r requirements.in
    #   pytest
isort==5.8.0
    # via
    #   -r requirements.in
    #   pylint
lazy-object-proxy==1.6.0
    # via
    #   -r requirements.in
    #   astroid
mccabe==0.6.1
    # via
    #   -r requirements.in
    #   pylint
packaging==20.9
    # via
    #   -r requirements.in
    #   pytest
pep517==0.10.0
    # via
    #   -r requirements.in
    #   pip-tools
pip-tools==6.1.0
    # via -r requirements.in
pluggy==0.13.1
    # via
    #   -r requirements.in
    #   pytest
py==1.10.0
    # via
    #   -r requirements.in
    #   pytest
pylint==2.8.3
    # via -r requirements.in
pyparsing==2.4.7
    # via
    #   -r requirements.in
    #   packaging
pytest-html==3.1.1
    # via -r requirements.in
pytest-metadata==1.11.0
    # via
    #   -r requirements.in
    #   pytest-html
pytest-repeat==0.9.1
    # via -r requirements.in
pytest==6.2.4
    # via
    #   -r requirements.in
    #   pytest-html
    #   pytest-metadata
    #   pytest-repeat
selenium==3.141.0
    # via
    #   -r requirements.in
    #   appium-python-client
toml==0.10.2
    # via
    #   -r requirements.in
    #   pep517
    #   pylint
    #   pytest
urllib3==1.26.5
    # via
    #   -r requirements.in
    #   selenium
wrapt==1.12.1
    # via
    #   -r requirements.in
    #   astroid

# The following packages are considered to be unsafe in a requirements file:
# pip
